Relocate all test packet sender code to udpHelpers.
