+++ server/udp/preloadGUI.js
@@ -3,24 +3,9 @@
 const bin = require('./udp/binary');
 const commConfig = require('../config/commConfig.js');
 var dgram = require('dgram');
-const makeSafetyUDP = require('../udp/helpers.js').makeSafetyUDP;
+const udpHelpers = require('../udp/helpers.js');
 
 module.exports = function () {
-  function sendPacket (sequence, type, payload, port) {
-    var testPacket = makeSafetyUDP(sequence, type, payload);
-    var packetBuf = new Buffer(testPacket);
-    var client = dgram.createSocket({type: 'udp4', reuseAddr: true});
-    client.bind();
-    client.on('listening', function () {
-      client.setBroadcast(true);
-      client.send(packetBuf, 0, packetBuf.length, port, commConfig.testDataGeneratorTargetHost, function (_err, bytes) {
-        client.close();
-      });
-    });
-
-    return testPacket;
-  }
-
   function preloadPowerNodes () {
     var BMSStreaming = [];
     BMSStreaming.push.apply(BMSStreaming, bin.uint32ToBytes(0, true)); // fault flags
@@ -77,8 +62,8 @@ module.exports = function () {
     BMSStreaming.push.apply(BMSStreaming, bin.uint32ToBytes(0, true)); // Temp Scan Count
     BMSStreaming.push.apply(BMSStreaming, bin.float32ToBytes(0, true)); // Pack Current
 
-    sendPacket(0, 0x3401, BMSStreaming, 9110);
-    sendPacket(0, 0x3401, BMSStreaming, 9111);
+    udpHelpers.sendPacket(0, 0x3401, BMSStreaming, 9110);
+    udpHelpers.sendPacket(0, 0x3401, BMSStreaming, 9111);
   }
 
   function GUIPreload (rtDataStore) {
